Move tags to their dedicated folder
